// src/types/apps/chatTypes.ts
export type StatusObjType = {
  busy: ThemeColor
  away: ThemeColor
  online: ThemeColor
  offline: ThemeColor
}
// ...
export type UserChatType = {
  _id: string
  firstName: string
  username: string
  lastName: string
  gender: string
  avatar: string | null
  role: string
  isActive: boolean
}

export type LastMessageType = {
  _id: string
  user: string
  content: string
  reply_to: string
  isDelete: boolean
  images: string[]
  createdAt: string
  updatedAt: string
}

export type ChatType = {
  _id?: string
  createdAt?: string
  updatedAt?: string
  user: UserChatType
  lastMessage?: LastMessageType
  unReadMessages: number
}

export type SelectedChatType = null | {
  chat: ChatsObj
  contact: ChatType
}

export type ChatStoreType = {
  chats: ChatType[] | null
  contacts: ContactType[] | null
  userProfile: ProfileUserType | null
  selectedChat: SelectedChatType
}

export type SendMsgParamsType = {
  chat?: ChatsObj
  message: string
  contact?: ChatType
}

export type ChatContentType = {
  hidden: boolean
  mdAbove: boolean
  user: UsersType | null
  sidebarWidth: number
  dispatch: Dispatch<any>
  statusObj: StatusObjType
  userProfileRightOpen: boolean
  selectedChat: ChatType | null
  setSelectedChat: React.Dispatch<React.SetStateAction<null | ChatType>>
  handleLeftSidebarToggle: () => void
  getInitials: (val: string) => string
  handleUserProfileRightSidebarToggle: () => void
  handleNewContact: () => void
}
// ...
export type MessageType = {
  _id: string
  user: UserChatType
  content: string
  reply_to: string
  isDelete: false
  images: string[]
  reacts: { [id: string]: string }
  seen: { [id: string]: string }
  createdAt: string
  updatedAt: string
}